\name{calculateApproximatePowerForDesignList}
\alias{calculateApproximatePowerForDesignList}
\title{calculateApproximatePowerForDesignList}
\usage{
calculateApproximatePowerForDesignList(designList, output.data.dir = ".")
}
\arguments{
  \item{designList}{list of pairs of design.glmmFG and glh
  objects}
}
\value{
design.glmmFG object with a single covariate
}
\description{
Calculate approximate power values for a list of design/glh
pairs.
}
\note{
This function requires several hours to run
}

