\name{calculateEmpiricalPowerForDesignList}
\alias{calculateEmpiricalPowerForDesignList}
\title{calculateEmpiricalPowerForDesignList}
\usage{
calculateEmpiricalPowerForDesignList(designList, output.data.dir = ".")
}
\arguments{
  \item{glh}{the general linear hypothesis object}

  \item{newColumns}{the updated number of columns}
}
\value{
writes the empirical power values as a CSV
}
\description{
Calculate empirical power values for a list of designs.
}
\note{
This function requires several hours to run
}

