% Generated by roxygen2 (4.0.2): do not edit by hand
\name{isPosDefSigmaE}
\alias{isPosDefSigmaE}
\title{isPosDefSigmaE}
\usage{
isPosDefSigmaE(designList)
}
\arguments{
\item{list}{of pairs of design.glmmFG and glh objects}
}
\value{
list of indices for designs with invalid residual error covariance
}
\description{
Convenience routine for checking if the designs in a
list have a valid residual error matrix
}
\keyword{internal}

