\name{runElashoffExample}
\alias{runElashoffExample}
\title{runElashoffExample}
\usage{
runElashoffExample()
}
\arguments{
  \item{study.seed}{the random number seed (defaults to
  7634)}

  \item{study.data.dir}{the directory into which data files
  are written (defaults to current working directory)}

  \item{study.figures.dir}{the directory into which pdf
  figures are written (defaults to the current working
  directory)}

  \item{study.runEmpirical}{if true, empirical power values
  will be recalculated. If false, existing empirical power
  values will be loaded from the R package.}
}
\description{
This function reproduces the applied example results for
the manuscript:\cr Kreidler, S. M., Muller, K. E., &
Glueck, D. H. Calculating Power for the General Linear
Multivariate Model With One or More Gaussian Covariates, In
review.
}
\note{
The example is based on the study by

Elashoff, D., Zhou, H., Reiss, J., Wang, J., Xiao, H.,
Henson, B., … Wong, D. T. W. (2012). Prevalidation of
salivary biomarkers for oral cancer detection. Cancer
Epidemiology, Biomarkers & Prevention, 21(4), 664–672.
}

