% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{simulateXMatrix}
\alias{simulateXMatrix}
\alias{simulateXMatrix,design.glmmFG,design.glmmFG-method}
\alias{simulateXMatrix,design.glmmFG-method}
\title{simulateXMatrix}
\usage{
simulateXMatrix(design)

\S4method{simulateXMatrix}{design.glmmFG}(design)
}
\arguments{
\item{design}{An object describing a study design, commonly a \code{design.glmmFG}}
}
\value{
A complete design matrix
}
\description{
Generate a complete X matrix for a given design.  If random
covariates are present in the design, the values of the covariates
will be randomly generated.
}
\seealso{
\code{\link{design.glmmFG}}
}

