\name{summarizeResults}
\alias{summarizeResults}
\title{summarizeResults}
\usage{
summarizeResults(output.data.dir = ".", output.figures.dir = ".")
}
\arguments{
  \item{designList}{list of pairs of design.glmmFG and glh
  objects}
}
\value{
design.glmmFG object with a single covariate
}
\description{
Generate box plots which summarize the deviations from
empirical power for the approximate methods.
}
\note{
This function requires several hours to run
}

