% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summarizeResults}
\alias{summarizeResults}
\title{summarizeResults}
\usage{
summarizeResults(output.data.dir = ".", output.figures.dir = ".")
}
\arguments{
\item{output.data.dir}{directory to which data sets are written}
}
\value{
output.figures.dir directory to which figures are written
}
\description{
Generate box plots which summarize the deviations from empirical
power for the approximate methods.
}

